Refactor: 가격 타입을 long로 수정

+++ src/main/java/com/capstone/rentit/item/domain/Item.java
@@ -45,7 +45,7 @@ public class Item {
     @Column(length = 500)
     private String damagedDescription;
 
-    private Integer price;
+    private long price;
 
     @Enumerated(EnumType.STRING)
     @Column(nullable = false)
